Extract image to scanned text code from `AddProductFromImageViewModel` to `ImageTextScanner`.
