+++ WooCommerce/Classes/ViewRelated/Products/Add Product/AddProductFromImage/AddProductFromImageViewModel.swift
@@ -46,12 +46,15 @@ final class AddProductFromImageViewModel: ObservableObject {
 
     private let siteID: Int64
     private let stores: StoresManager
+    private let imageTextScanner: ImageTextScannerProtocol
 
     init(siteID: Int64,
          stores: StoresManager = ServiceLocator.stores,
+         imageTextScanner: ImageTextScannerProtocol = ImageTextScanner(),
          onAddImage: @escaping (MediaPickingSource) async -> MediaPickerImage?) {
         self.siteID = siteID
         self.stores = stores
+        self.imageTextScanner = imageTextScanner
         self.onAddImage = onAddImage
 
         selectedImageSubscription = $imageState.compactMap { $0.image?.image }
@@ -85,51 +88,15 @@ final class AddProductFromImageViewModel: ObservableObject {
 
 private extension AddProductFromImageViewModel {
     func onSelectedImage(_ image: UIImage) {
-        // Gets the CGImage on which to perform requests.
-        guard let cgImage = image.cgImage else {
-            return
-        }
-
-        // Creates a new image-request handler.
-        let requestHandler = VNImageRequestHandler(cgImage: cgImage)
-
-        // Creates a new request to recognize text.
-        let request = VNRecognizeTextRequest { [weak self] request, error in
-            self?.onScannedTextRequestCompletion(request: request, error: error)
-        }
-
-        if #available(iOS 16.0, *) {
-            request.revision = VNRecognizeTextRequestRevision3
-            request.automaticallyDetectsLanguage = true
-        }
-
-        do {
-            // Performs the text-recognition request.
-            try requestHandler.perform([request])
-        } catch {
-            DDLogError("⛔️ Unable to perform image text generation request: \(error)")
-        }
-    }
-
-    func onScannedTextRequestCompletion(request: VNRequest, error: Error?) {
-        let texts = scannedTexts(from: request)
-        scannedTexts = texts.map { .init(text: $0, isSelected: true) }
         Task { @MainActor in
-            isGeneratingDetails = true
-            await generateAndPopulateProductDetails(from: texts)
-            isGeneratingDetails = false
+            do {
+                let texts = try await imageTextScanner.scanText(from: image)
+                scannedTexts = texts.map { .init(text: $0, isSelected: true) }
+                generateProductDetails()
+            } catch {
+                DDLogError("⛔️ Error scanning text from image: \(error)")
+            }
         }
-    }
-
-    func scannedTexts(from request: VNRequest) -> [String] {
-        guard let observations = request.results as? [VNRecognizedTextObservation] else {
-            return []
-        }
-        let recognizedStrings = observations.compactMap { observation in
-            // Returns the string of the top VNRecognizedText instance.
-            observation.topCandidates(1).first?.string
-        }
-        return recognizedStrings
     }
 
     func generateAndPopulateProductDetails(from scannedTexts: [String]) async {
